Task5_Airflow/requirements.txt
apache-airflow-providers-mongo>=1.0.0
pandas>=2.0.0
pymongo>=4.0.0
dnspython>=2.0.0